src/modelgauge/default_annotator_set.py
from typing import Any, Dict, Optional

from modelgauge.annotator import Annotator
from modelgauge.annotator_set import AnnotatorSet
from modelgauge.annotators.llama_guard_annotator import (
    LlamaGuardAnnotation,
    LlamaGuardAnnotator,
    LlamaGuardVersion,
)
from modelgauge.single_turn_prompt_response import (
    TestItemAnnotations,
)
from modelgauge.suts.together_client import TogetherApiKey


class DefaultAnnotatorSet(AnnotatorSet):

    ANNOTATOR_NAME = "llama_guard_2"
    annotators: dict[str, Annotator] = {}
    secrets: dict[str, Any] = {}

    required_secret_types = {
        "together_api_key": TogetherApiKey,
    }

    def __init__(self, secrets):
        self.secrets = secrets
        self._load_annotators()

    def _load_annotators(self):
        self.annotators[DefaultAnnotatorSet.ANNOTATOR_NAME] = LlamaGuardAnnotator(
            DefaultAnnotatorSet.ANNOTATOR_NAME,
            self.secrets["together_api_key"],
            llama_guard_version=LlamaGuardVersion.VERSION_2,
        )

    def evaluate(self, item: TestItemAnnotations) -> Dict[str, float]:
        annotation = (
            item.interactions[0]
            .response.completions[0]
            .get_annotation(DefaultAnnotatorSet.ANNOTATOR_NAME, LlamaGuardAnnotation)
        )
        return {"is_safe": float(annotation.is_safe)}
